Changes to the FlywayMigrationStandalone to work with environment variables only
