+++ eval/ticketExchange/src/main/java/br/com/delogic/ticketExchange/controller/EventController.java
@@ -1,4 +1,4 @@
-package br.com.delogic.ticketExchange.controler;
+package br.com.delogic.ticketExchange.controller;
 
 import java.util.Optional;
 
@@ -10,30 +10,30 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RestController;
 
-import br.com.delogic.ticketExchange.dto.UserDTO;
-import br.com.delogic.ticketExchange.service.UserService;
+import br.com.delogic.ticketExchange.dto.EventDTO;
+import br.com.delogic.ticketExchange.service.EventService;
+
 
 @RestController
-public class UserController extends RootController{
+public class EventController extends RootController{
 	
 	@Autowired
-    private UserService userService;
+    private EventService eventService;
 
         
-    @GetMapping("/users")
-    public ResponseEntity<Page<Long>> getAllUserIds(Pageable pageable) {    	
-    	Optional<Page<Long>> result = userService.getAllIds(pageable);
+    @GetMapping("/events")
+    public ResponseEntity<Page<Long>> getAllEventIds(Pageable pageable) {    	
+    	Optional<Page<Long>> result = eventService.getAllIds(pageable);
     	return result.map(ResponseEntity::ok)
                      .orElse(ResponseEntity.notFound().build());    
     }
     
 
-    @GetMapping("/users/{id}")
-    public ResponseEntity<UserDTO> getUserById(@PathVariable Long id) {
-        return userService.getUserById(id)
+    @GetMapping("/events/{id}")
+    public ResponseEntity<EventDTO> getEventById(@PathVariable Long id) {
+        return eventService.getEventById(id)
                 .map(ResponseEntity::ok)
                 .orElse(ResponseEntity.notFound().build());
     }
-	
 
 }
