+++ eval/ticketExchange/src/main/java/br/com/delogic/ticketExchange/controller/PromotionalControler.java
@@ -1,4 +1,4 @@
-package br.com.delogic.ticketExchange.controler;
+package br.com.delogic.ticketExchange.controller;
 
 import java.util.Date;
 import java.util.List;
