+++ eval/ticketExchange/src/main/java/br/com/delogic/ticketExchange/controller/UserController.java
@@ -1,4 +1,4 @@
-package br.com.delogic.ticketExchange.controler;
+package br.com.delogic.ticketExchange.controller;
 
 import java.util.Optional;
 
@@ -10,30 +10,30 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RestController;
 
-import br.com.delogic.ticketExchange.dto.ListingDTO;
-import br.com.delogic.ticketExchange.service.ListingService;
-
+import br.com.delogic.ticketExchange.dto.UserDTO;
+import br.com.delogic.ticketExchange.service.UserService;
 
 @RestController
-public class ListingController extends RootController{
+public class UserController extends RootController{
 	
 	@Autowired
-    private ListingService listingService;
+    private UserService userService;
 
         
-    @GetMapping("/listings")
-    public ResponseEntity<Page<Long>> getAllListingIds(Pageable pageable) {    	
-    	Optional<Page<Long>> result = listingService.getAllIds(pageable);
+    @GetMapping("/users")
+    public ResponseEntity<Page<Long>> getAllUserIds(Pageable pageable) {    	
+    	Optional<Page<Long>> result = userService.getAllIds(pageable);
     	return result.map(ResponseEntity::ok)
                      .orElse(ResponseEntity.notFound().build());    
     }
     
 
-    @GetMapping("/listings/{id}")
-    public ResponseEntity<ListingDTO> getListingById(@PathVariable Long id) {
-        return listingService.getListingById(id)
+    @GetMapping("/users/{id}")
+    public ResponseEntity<UserDTO> getUserById(@PathVariable Long id) {
+        return userService.getUserById(id)
                 .map(ResponseEntity::ok)
                 .orElse(ResponseEntity.notFound().build());
     }
+	
 
 }
