+++ eval/ticketExchange/src/main/java/br/com/delogic/ticketExchange/migration/flyway/FlywayMigrationStandalone.java
@@ -1,13 +1,7 @@
 package br.com.delogic.ticketExchange.migration.flyway;
 
-import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.util.Properties;
-
-import javax.sql.DataSource;
 
 import org.flywaydb.core.Flyway;
 
@@ -18,10 +12,7 @@ import org.flywaydb.core.Flyway;
 
 public class FlywayMigrationStandalone {
 	
-	final static String PROPERTIES_FILE =  "src/main/resources/application.properties";  
-	final static String TEST_PROPERTIES_FILE =  "src/main/resources/application-test.properties";
-	
-	
+		
     public static void main(String[] args) throws IOException {
      	
         doMigration();    
@@ -30,22 +21,16 @@ public class FlywayMigrationStandalone {
     
     public static void doMigration() throws FileNotFoundException, IOException {
     	
-    	doBaseMigration(PROPERTIES_FILE);
-    	
+    	doBaseMigration();    	
     }
 
 
-	private static void doBaseMigration(String propertiesFileName) throws IOException, FileNotFoundException {
-		Properties properties = new Properties();
-
-        try (FileInputStream fis = new FileInputStream(propertiesFileName)) {
-            // Load the property file
-            properties.load(fis);
-
-            // Retrieve properties using their keys
-            String username = properties.getProperty("spring.datasource.username");
-            String password = properties.getProperty("spring.datasource.password");
-            String url = properties.getProperty("spring.datasource.url");
+	private static void doBaseMigration(){
+      
+                    
+        	String url = System.getenv("DATABASE_URL");
+        	String username = System.getenv("DATABASE_USER");
+        	String password = System.getenv("DATABASE_PASSWORD");      	
                        
  
             // Print the retrieved properties            
@@ -62,6 +47,6 @@ public class FlywayMigrationStandalone {
 
             System.out.println("All Migrations executed.");
             
-        }
+        
 	}
 }
