+++ eval/ticketExchange/src/main/resources/application-test.properties
@@ -7,5 +7,6 @@ spring.datasource.driver-class-name=org.h2.Driver
 spring.datasource.driverClassName=org.h2.Driver
 spring.datasource.username=sa
 spring.datasource.password=
+
 spring.jpa.hibernate.ddl-auto=validate
