+++ eval/ticketExchange/src/main/resources/application.properties
@@ -1,17 +1,17 @@
 spring.application.name=ticketExchange
 
+spring.datasource.url=${DATABASE_URL}
+spring.datasource.username=${DATABASE_USER}
+spring.datasource.password=${DATABASE_PASSWORD}
 
-
-spring.datasource.url=jdbc:mysql://localhost:3307/delogic
-spring.datasource.username=root
-spring.datasource.password=root
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.datasource.driverClassName=com.mysql.cj.jdbc.Driver
+
+##spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL8Dialect
 
 
 spring.jpa.show-sql=true 
 logging.level.org.springframework=DEBUG
-
 
 
 spring.flyway.enabled=false
