+++ eval/ticketExchange/src/test/java/br/com/delogic/ticketExchange/controller/CategoryControllerTest.java
@@ -18,7 +18,7 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
 
-import br.com.delogic.ticketExchange.controler.CategoryController;
+import br.com.delogic.ticketExchange.controller.CategoryController;
 import br.com.delogic.ticketExchange.dto.CategoryDTO;
 import br.com.delogic.ticketExchange.service.CategoryService;
