+++ eval/ticketExchange/src/test/java/br/com/delogic/ticketExchange/controller/DateControllerTest.java
@@ -18,7 +18,7 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
 
-import br.com.delogic.ticketExchange.controler.DateController;
+import br.com.delogic.ticketExchange.controller.DateController;
 import br.com.delogic.ticketExchange.dto.DateDTO;
 import br.com.delogic.ticketExchange.service.DateService;
