+++ eval/ticketExchange/src/test/java/br/com/delogic/ticketExchange/controller/SaleControllerTest.java
@@ -18,7 +18,7 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
 
-import br.com.delogic.ticketExchange.controler.SalesController;
+import br.com.delogic.ticketExchange.controller.SalesController;
 import br.com.delogic.ticketExchange.dto.SaleDTO;
 import br.com.delogic.ticketExchange.service.SaleService;
